Move the `isMavenCentralUrl` function from `core-utils` to `common-utils`
This function can be used independently of ORT.
