+++ utils/common/src/main/kotlin/Utils.kt
@@ -68,6 +68,13 @@ fun getCommonFileParent(files: Collection<File>): File? =
         if (commonPrefix.isDirectory) commonPrefix else commonPrefix.parentFile
     }
 
+private val mavenCentralUrlPattern = Regex("^https?://repo1?\\.maven(\\.apache)?\\.org(/.*)?$")
+
+/**
+ * Return whether the given [url] points to Maven Central or not.
+ */
+fun isMavenCentralUrl(url: String) = url.matches(mavenCentralUrlPattern)
+
 /**
  * Return the concatenated [strings] separated by [separator] whereas blank strings are omitted.
  */
