+++ utils/common/src/test/kotlin/UtilsTest.kt
@@ -21,6 +21,7 @@
 package org.ossreviewtoolkit.utils.common
 
 import io.kotest.core.spec.style.WordSpec
+import io.kotest.inspectors.forAll
 import io.kotest.matchers.collections.containExactly
 import io.kotest.matchers.nulls.beNull
 import io.kotest.matchers.should
@@ -55,6 +56,30 @@ class UtilsTest : WordSpec({
                 "/a",
                 "/"
             )
+        }
+    }
+
+    "isMavenCentralUrl" should {
+        "return true for URLs that point to Maven Central" {
+            listOf(
+                "https://repo.maven.apache.org/maven2",
+                "https://repo.maven.apache.org",
+                "http://repo.maven.apache.org",
+                "https://repo1.maven.org/maven2",
+                "https://repo1.maven.org",
+                "http://repo1.maven.org"
+            ).forAll {
+                isMavenCentralUrl(it) shouldBe true
+            }
+        }
+
+        "return false for URLs that do not point to Maven Central" {
+            listOf(
+                "https://repo2.maven.org",
+                "https://github.com"
+            ).forAll {
+                isMavenCentralUrl(it) shouldBe false
+            }
         }
     }
 })
