+++ utils/core/src/main/kotlin/Utils.kt
@@ -27,8 +27,6 @@ import org.ossreviewtoolkit.utils.common.toUri
 import org.ossreviewtoolkit.utils.common.withoutPrefix
 import org.ossreviewtoolkit.utils.common.withoutSuffix
 
-private val mavenCentralUrlPattern = Regex("^https?://repo1?\\.maven(\\.apache)?\\.org(/.*)?$")
-
 /**
  * The directory to store ORT (read-only) configuration in.
  */
@@ -141,11 +139,6 @@ fun installAuthenticatorAndProxySelector(): OrtProxySelector {
     OrtAuthenticator.install()
     return OrtProxySelector.install()
 }
-
-/**
- * Return whether the given [url] points to Maven Central or not.
- */
-fun isMavenCentralUrl(url: String) = url.matches(mavenCentralUrlPattern)
 
 /**
  * Normalize a string representing a [VCS URL][vcsUrl] to a common string form.
